feat!: Change expiration from usize to u64

+++ common/src/models/rest/request/create_short_request.rs
@@ -3,11 +3,11 @@ use serde::{Deserialize, Serialize};
 #[derive(Clone, Debug, Deserialize, Serialize)]
 pub struct CreateShortRequest {
     pub long_url: String,
-    pub expiration: Option<usize>,
+    pub expiration: Option<u64>,
 }
 
 impl CreateShortRequest {
-    pub fn new(long_url: String, expiration: Option<usize>) -> CreateShortRequest {
+    pub fn new(long_url: String, expiration: Option<u64>) -> CreateShortRequest {
         CreateShortRequest {
             long_url,
             expiration,
